requirements.txt
Click==7.0
numba==0.48.0
numpy==1.17.4
tqdm==4.30.0
cartopy==0.19.0.post1
pandas==1.2.4
Pillow==8.2.0
pytorch_lightning==1.3.2
torch==1.8.1
torchvision==0.9.1
